Link answers and annotations to users
